loading screen, progress bar is done but does not properly start next activity
