+++ app/src/main/java/org/danielsoares/pickupapp/MainActivity.java
@@ -1,31 +1,36 @@
 package org.danielsoares.pickupapp;
 
-import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.view.View;
-import android.widget.EditText;
+import android.view.WindowManager;
+import android.widget.ProgressBar;
 import android.widget.TextView;
 
 public class MainActivity extends AppCompatActivity {
 
-    private TextView clickThisForNow;
+    ProgressBar progressBar;
+    TextView textView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_loading_screen);
 
-        clickThisForNow = findViewById(R.id.clickThisForNow);
+        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
+                WindowManager.LayoutParams.FLAG_FULLSCREEN);
 
-        // Directs you to login page
-        clickThisForNow.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent forNowIntent = new Intent(MainActivity.this, Login.class);
-                MainActivity.this.startActivity(forNowIntent);
-            }
+        progressBar = findViewById(R.id.progress_bar);
+        textView = findViewById(R.id.text_view);
 
-        });
+        progressBar.setMax(100);
+        progressBar.setScaleY(3f);
+
+        progressAnimation();
+    }
+
+    public void progressAnimation() {
+        ProgressBarAnimation anim = new ProgressBarAnimation( MainActivity.this, progressBar, textView, 0f, 100f);
+        anim.setDuration(8000);
+        progressBar.setAnimation(anim);
     }
 }
