--- app/src/main/java/org/danielsoares/pickupapp/TestHomeActivity.java
@@ -1,32 +1,0 @@
-package org.danielsoares.pickupapp;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.widget.ProgressBar;
-import android.widget.TextView;
-
-public class TestHomeActivity extends AppCompatActivity {
-
-    ProgressBar progressBar;
-    TextView textView;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_test_home);
-
-        progressBar = findViewById(R.id.progress_bar);
-        textView = findViewById(R.id.text_view);
-
-        progressBar.setMax(100);
-        progressBar.setScaleY(3f);
-
-        progressAnimation();
-    }
-
-    public void progressAnimation() {
-        ProgressBarAnimation anim = new ProgressBarAnimation( this, progressBar, textView, 0f, 100f);
-        anim.setDuration(8000);
-        progressBar.setAnimation(anim);
-    }
-}
