--- app/src/main/res/layout/content_loading_screen.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="16dp"
-    android:background="@drawable/loadingscreen"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".LoadingScreen"
-    tools:showIn="@layout/activity_loading_screen">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="0 %"
-        android:textColor="#fff"
-        android:textSize="25sp"
-        android:id="@+id/text_view"
-        android:layout_centerHorizontal="true"
-        android:layout_above="@+id/progress_bar"
-        />
-
-    <ProgressBar
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/progress_bar"
-        style="@style/Widget.AppCompat.ProgressBar.Horizontal"
-        android:layout_alignParentBottom="true"
-        android:layout_marginBottom="20dp"
-    />
-
-</RelativeLayout>
-
